Graceful shutdown på skyggesak og saksstatistikkscheduler
